Fix: earlier keys take priority in ALIST-HASH-TABLE and PLIST-HASH-TABLE
Also record tests for the new behavior.

This is a semantic change, but more in line with how alists and
plists are defined in the CLHS.

See '15.6. Association Lists' and 'property list' in the Glossary.

Reported by Christoph Arenz to alexandria-devel in
'Wrong handling of duplicate keys in alist-hash-table and plist-hash-table'.